requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.41.2
aiohttp==3.9.3
boto3==1.34.34
Pillow==10.2.0
python-dotenv==1.0.1
pymongo==4.6.1 